feat(bitbucket-server): add rate limiting option
This commit introduces an optional `rateLimit` property to the
`BitbucketServerIntegrationConfig`. When provided, the `rateLimit`
determines how long to wait in milliseconds between calls to the
Bitbucket Server API. The default is `undefined`, which means no rate
limiting will be applied. The 'bottleneck' package is added to manage
the rate limiting process.

Signed-off-by: Kevin Johnson <kljohnson@verisk.com>